+++ src/sdk.ts
@@ -2,22 +2,14 @@ import Popup from './authentication/popup';
 import Redirect from './authentication/redirect';
 import Iframe from './authentication/iframe';
 import Transaction from './authentication/transaction';
-import {
-  type TransactionParams,
-  type TransactionData,
-  type VerifyInput,
-} from './types/transaction';
+import {type TransactionData, type VerifyInput} from './types/transaction';
 import qs from 'qs';
 import sjcl from './vendor/sjcl';
-import {pick} from './helpers/object';
+import {pick, deepMerge, omitBy} from './helpers/object';
 import encoding from './helpers/encoding';
 import RedirectUriParamsPersister from './helpers/persisters/redirectUriParams';
 import random from './helpers/random';
-import {
-  type SDKOptions,
-  type ResolvedOptions,
-} from './types/sdk';
-import {type PersistParams} from './types/persister';
+import {type SDKOptions, type ResolvedOptions} from './types/sdk';
 import {
   type PopupSDK,
   type IframeSDK,
@@ -37,7 +29,8 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
       throw new Error('client id not provided');
     }
 
-    const defaultOptions: Omit<ResolvedOptions, 'client_id'> = {
+    const defaultOptions: ResolvedOptions = {
+      client_id: '',
       organization_id: '',
       prompt: '',
       response_type: 'token',
@@ -65,7 +58,7 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
       },
     };
 
-    this.options = Object.assign({}, defaultOptions, options);
+    this.options = deepMerge(defaultOptions, options);
     this.transaction = new Transaction(this.options);
     this.redirectUriParamsPersister = new RedirectUriParamsPersister(
       this.options,
@@ -76,7 +69,7 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
    * Use iframe for authorization. Not recommended due to ITP 2.0.
    */
   iframe(options: Partial<SDKOptions> = {}): Iframe {
-    const localOptions = Object.assign({}, this.options, options);
+    const localOptions = deepMerge(this.options, options);
     return new Iframe(this, localOptions);
   }
 
@@ -84,7 +77,7 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
    * Use popup for authorization. Must be called inside a click handler.
    */
   popup(options: Partial<SDKOptions> = {}): Popup {
-    const localOptions = Object.assign({}, this.options, options);
+    const localOptions = deepMerge(this.options, options);
     return new Popup(this, localOptions);
   }
 
@@ -92,7 +85,7 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
    * Use redirect for authorization.
    */
   redirect(options: Partial<SDKOptions> = {}): Redirect {
-    const localOptions = Object.assign({}, this.options, options);
+    const localOptions = deepMerge(this.options, options);
     return new Redirect(this, localOptions);
   }
 
@@ -102,11 +95,8 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
    * @param flow - Set to `'button'` for popup and iframe flows.
    */
   authorizeURL(options: Partial<SDKOptions> = {}, flow = ''): string {
-    const localOptions: ResolvedOptions = Object.assign(
-      {},
-      this.options,
-      options,
-    );
+    const localOptions: ResolvedOptions = deepMerge(this.options, options);
+
     if (!localOptions.state) {
       localOptions.state = random.string(localOptions.transaction.key_length);
     }
@@ -115,15 +105,19 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
       localOptions.redirect_uri = window.location.href;
     }
 
-    const params: Partial<ResolvedOptions> = pick(localOptions, [
-      'client_id',
-      'organization_id',
-      'redirect_uri',
-      'state',
-      'response_type',
-      'scope',
-      'prompt',
-    ]);
+    const params: Partial<ResolvedOptions> &
+      Pick<ResolvedOptions, 'redirect_uri' | 'state' | 'code_verifier'> = pick(
+      localOptions,
+      [
+        'client_id',
+        'organization_id',
+        'redirect_uri',
+        'state',
+        'response_type',
+        'scope',
+        'prompt',
+      ],
+    );
 
     Object.assign(params, localOptions.tracking);
 
@@ -184,12 +178,17 @@ export default class AccountsSDK implements PopupSDK, IframeSDK, RedirectSDK {
       }
     }
 
-    this.transaction.generate(params as TransactionParams);
-    this.redirectUriParamsPersister.persist(params as PersistParams);
+    this.transaction.generate(params);
+    this.redirectUriParamsPersister.persist(params);
 
     delete params.code_verifier;
 
-    return url + '?' + qs.stringify(params);
+    const cleanedParams = omitBy(
+      params,
+      (value) => value === '' || value === null,
+    );
+
+    return url + '?' + qs.stringify(cleanedParams);
   }
 
   /**